Removed Guzzle as an explicit Composer dependency.
There was a need for a lightweight version of this library that uses guzzle.phar instead of the full Guzzle source code. This takes the size of the library down from more than 11Mb to around 1Mb.

If a project uses this library and it already has Guzzle as a direct dependency, guzzle.phar will not be loaded.